\name{epmgp-package}
\alias{epmgp-package}
\alias{epmgp}
\docType{package}
\title{
\packageTitle{epmgp}
}
\description{
\packageDescription{epmgp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{epmgp}
\packageIndices{epmgp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{epmgp}

Maintainer: \packageMaintainer{epmgp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
